+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## next version
 
 + NewJieba(...string) support variable arguments
++ NewExtractor(...string) support variable arguments
 + removed example/ and write Example in `*_test.go`
 + add some kind of Benchmark, Testing
 + fix bug in extractor.go: calling C.free manully to free momery allocated by C.CString